Extract additional utilities
